+++ tests/test/kotlin/com/rian/osu/beatmap/UnadjustedStandardHitWindowTest.kt
@@ -6,27 +6,27 @@ import org.junit.Test
 class UnadjustedStandardHitWindowTest : HitWindowTest() {
     @Test
     fun `Test hit window`() {
-        testHitWindowValues(10f, 20f, 60f, 100f)
-        testHitWindowValues(8.2f, 30.8f, 74.4f, 118f)
-        testHitWindowValues(6.5f, 41f, 88f, 135f)
-        testHitWindowValues(3.7f, 57.8f, 110.4f, 163f)
-        testHitWindowValues(-1.6f, 89.6f, 152.8f, 216f)
+        testHitWindowValues(10.0, 20.0, 60.0, 100.0)
+        testHitWindowValues(8.2, 30.8, 74.4, 118.0)
+        testHitWindowValues(6.5, 41.0, 88.0, 135.0)
+        testHitWindowValues(3.7, 57.8, 110.4, 163.0)
+        testHitWindowValues(-1.6, 89.6, 152.8, 216.0)
     }
 
     @Test
     fun `Test hit window to OD conversion`() {
-        fun testConversion(od: Float, greatWindow: Float, okWindow: Float, mehWindow: Float) {
-            Assert.assertEquals(od, UnadjustedStandardHitWindow.hitWindow300ToOverallDifficulty(greatWindow), 1e-2f)
-            Assert.assertEquals(od, UnadjustedStandardHitWindow.hitWindow100ToOverallDifficulty(okWindow), 1e-2f)
-            Assert.assertEquals(od, UnadjustedStandardHitWindow.hitWindow50ToOverallDifficulty(mehWindow), 1e-2f)
+        fun testConversion(od: Double, greatWindow: Double, okWindow: Double, mehWindow: Double) {
+            Assert.assertEquals(od, UnadjustedStandardHitWindow.hitWindow300ToOverallDifficulty(greatWindow), 1e-2)
+            Assert.assertEquals(od, UnadjustedStandardHitWindow.hitWindow100ToOverallDifficulty(okWindow), 1e-2)
+            Assert.assertEquals(od, UnadjustedStandardHitWindow.hitWindow50ToOverallDifficulty(mehWindow), 1e-2)
         }
 
-        testConversion(10f, 20f, 60f, 100f)
-        testConversion(8.2f, 30.8f, 74.4f, 118f)
-        testConversion(6.5f, 41f, 88f, 135f)
-        testConversion(3.7f, 57.8f, 110.4f, 163f)
-        testConversion(-1.6f, 89.6f, 152.8f, 216f)
+        testConversion(10.0, 20.0, 60.0, 100.0)
+        testConversion(8.2, 30.8, 74.4, 118.0)
+        testConversion(6.5, 41.0, 88.0, 135.0)
+        testConversion(3.7, 57.8, 110.4, 163.0)
+        testConversion(-1.6, 89.6, 152.8, 216.0)
     }
 
-    override fun createHitWindow(od: Float) = UnadjustedStandardHitWindow(od)
+    override fun createHitWindow(od: Double) = UnadjustedStandardHitWindow(od)
 }
